Moved Replicator class to its own file + draft for Replicator test
