+++ testsuite/libunderpass.all/rep-test.cc
@@ -19,106 +19,40 @@
 
 #include <dejagnu.h>
 #include <iostream>
-#include <string>
-#include <pqxx/pqxx>
-
-#include "hottm.hh"
-#include "galaxy/galaxy.hh"
-
-#include <boost/date_time.hpp>
+#include "log.hh"
 #include "boost/date_time/posix_time/posix_time.hpp"
-#include "boost/date_time/gregorian/gregorian.hpp"
+#include "galaxy/osmchange.hh"
+#include "replicatorconfig.hh"
+#include "galaxy/replicator.hh"
 
-#include "galaxy/replication.hh"
-#include "data/osmobjects.hh"
-
+using namespace logger;
+using namespace replicatorconfig;
+using namespace replicator;
 using namespace boost::posix_time;
-using namespace boost::gregorian;
-using namespace galaxy;
-
-TestState runtest;
-
-class TestRep : public replication::Planet
-{
-public:
-    TestRep(void) {
-    };
-};
 
 int
-main(int argc, char *argv[])
-{
+main(int argc, char *argv[]) {
 
-    TestRep tr;
+    logger::LogFile &dbglogfile = logger::LogFile::getDefaultInstance();
+    dbglogfile.setWriteDisk(true);
+    dbglogfile.setLogFilename("rep-test.log");
+    dbglogfile.setVerbosity(3);
 
-    // 000 tests
-    auto result = tr.findData(replication::changeset, time_from_string("2014-09-07 10:23"));
-    std::cout << result << std::endl;
-    if (result->path == "000/956/986") {
-        runtest.pass("Planet::findData(no state 000)");
-    } else {
-        runtest.fail("Planet::findData(no state 000)");
-    }
-    // auto foo = tr.getState(replication::changeset, time_from_string("2014-11-07 10:23"));
-    // if (foo->path == "000/962/746") {
-    //     runtest.pass("Planet::findData(no state 000)");
+    ReplicatorConfig config;
+    config.planet_server = config.planet_servers[0].domain + "/replication";
+    std::string ts("2020-01-01T00:00:00");
+    config.start_time = from_iso_extended_string(ts);
+
+    replicator::Replicator replicator;
+    auto osmchange = replicator.findRemotePath(config, config.start_time);
+    osmchange->dump();
+
+    // if () {
+    //     runtest.pass("Find remote path from timestamp");    
     // } else {
-    //     runtest.fail("Planet::findData(no state 000)");
+    //     runtest.fail("Find remote path from timestamp");    
     // }
-
-    // 001 tests
-    result = tr.findData(replication::changeset, time_from_string("2016-08-04 09:03"));
-    if (result->path == "001/958/981") {
-        runtest.pass("Planet::findData(no state 001)");
-    } else {
-        runtest.fail("Planet::findData(no state 001)");
-    }
-
-    // 002 tests
-    result = tr.findData(replication::changeset, time_from_string("2017-09-07 18:22"));
-    if (result->path == "002/532/990") {
-        runtest.pass("Planet::findData(has state)");
-    } else {
-        runtest.fail("Planet::findData(has state)");
-    }
-
-    replication::RemoteURL remote("https://planet.openstreetmap.org/replication/changesets/000/956/986");
-    // remote.dump();
-    if (remote.minor == 956 && remote.index == 986) {
-        runtest.pass("RemoteURL(url)");
-    } else {
-        runtest.fail("RemoteURL(url)");
-    }
-    remote.Increment();
-    // remote.dump();
-    if (remote.minor == 956 && remote.index == 987) {
-        runtest.pass("RemoteURL.Increment()");
-    } else {
-        runtest.fail("RemoteURL.Increment()");
-    }
-    remote.parse("https://planet.openstreetmap.org/replication/changesets/000/956/999");
-    remote.Increment();
-    // remote.dump();
-    if (remote.minor == 957 && remote.index == 0) {
-        runtest.pass("RemoteURL.Increment(minor roll-over)");
-    } else {
-        runtest.fail("RemoteURL.Increment(minor roll-over)");
-    }
-    remote.parse("https://planet.openstreetmap.org/replication/changesets/000/999/999");
-    remote.Increment();
-    // remote.dump();
-    if (remote.major == 1 && remote.minor == 0 && remote.index == 1) {
-        runtest.pass("RemoteURL.Increment(major roll-over)");
-    } else {
-        runtest.fail("RemoteURL.Increment(major roll-over)");
-    }
     
-    // if (tr.checkTag("building", "yes") == true) {
-    //     runtest.pass("Validate::checkTag(good tag)");
-    // } else {
-    //     runtest.fail("Validate::checkTag(good tag)");
-    // }
-
 }
 
 // local Variables:
